+++ packages/base/src/components/Modal/Modal.types.ts
@@ -8,6 +8,8 @@ export const ModalSize = {
   L: 'L',
   XL: 'XL',
   XXL: 'XXL',
+  XXXL: 'XXXL',
+  FULL: 'FULL',
 } as const;
 
 export type ModalSizeType = $Values<typeof ModalSize>;