+++ packages/base/src/components/Popper/usePopper.ts
@@ -1,8 +1,9 @@
 import { createContext } from 'react';
 import * as React from 'react';
 import type { ReferenceType } from '@floating-ui/react';
+import type { UseFloatingReturn } from '@floating-ui/react-dom';
 
-type PopperContextType<RT extends ReferenceType = ReferenceType> = {
+type PopperContextType<RT extends ReferenceType = ReferenceType> = Omit<UseFloatingReturn<RT>, 'refs'> & {
   isOpen: boolean;
   setIsOpen: (isOpen: boolean) => void;
   reference: React.MutableRefObject<RT | null>;
@@ -11,6 +12,7 @@ type PopperContextType<RT extends ReferenceType = ReferenceType> = {
   setFloating: (node: HTMLElement | null) => void;
   floatingStyles: React.CSSProperties;
 };
+
 export const PopperContext = createContext<PopperContextType | null>(null);
 
 export const usePopper = () => {