+++ packages/utils/src/utils/index.ts
@@ -5,3 +5,4 @@ export * from './noop';
 export * from './throttle';
 export * from './isNullOrUndefined';
 export * from './OutSideClick';
+export * from './SwitchCase';